requirement.txt
request
random
rich
EbookLib
pyyaml
